Made changes to the Arena and added a beacon
I enlarged the arena to accomodate the number of robots required for our experiments.

I edited the distribute tag to make robots spawn in a grid across the arena from the beacon.

Robots per row and column can be changed by editing  the layout attribute of the the position tag inside the distribute tag. 

I left comments in the code indicating how to edit the layout attribute. The number of columns should be kept constant so that the distance between the edge of the swarm and the beacon remains constant too.

I also had to rebuild my project. I'm pushing those changes too.