Add files via upload
I started to implement the flocking function. I have part of the flocking function finished, which is to determine the long range attraction of epucks.This is located in the function flocking(float obstacle_avoidance_timer) which returns a coherence value that will be used to find the swarm of the robots. The function is missing the short range repulsion that will be combined with the return value of the long range attraction. The function is also missing code that will turn control the angle of the robot.