Add files via upload
Flocking and Control Step Method Updates:
Flocking method:
1. Renamed the original flocking method to "flocking_long_range" because this methods main focus is the long range attraction between the epuck's.
2. Added 4 different parameters, to the method keep track of the current x and y position of an epuck as well update the epuck's velocity and making use of its sensors.
3. Added new variables that are used for finding the x and y position of the swarms center as well as the velocity that is need to reach the center.
Control Step Method:
1. Deleted the code that we used to randomly move the epucks throughout.
2. Added a section at the method of the Control Step , its main use is to debug the flocking_long_range function
Currently the code runs and you can see the results of the updated code if you run argos with only 2 epucks the arena. One of the epucks will stay with in range of the "lead" epuck as it moves forward and across the grid. A couple of test still need to be made to test the effectiveness of the code so far, such as:
1. Test the code with more than 10 epucks.
2. Test if the flocking method isnt just hard coding the epuck to follow, this will be done by not only doing the above test but also making the lead epuck move randomly throughout the grid.