Reimplemented short range repulsion using range and bearing sensor
This is the first commit since branching off from master.

We decided to pivot our approach to short range repulsion by using the range and bearing sensor readings instead of the proximity sensor readings. This approach is much easier, effective, and more tunable.

Currently experiencing a bug that we have not seen for a while: 

[FATAL] CGrid<ENTITY>::PositionToCell() : Position <nan,nan,-nan> out of bounds X -> -3:3 Y -> -4:4 Z -> 0:1

Testing will follow...